+++ passthrough/usecase/passthrough_usecase.go
@@ -25,7 +25,9 @@ type passthroughUseCase struct {
 	liquidityPricer       domain.LiquidityPricer
 	passthroughGRPCClient passthroughdomain.PassthroughGRPCClient
 
-	aprPrefetcher datafetchers.Fetcher[[]passthroughdomain.PoolAPR]
+	aprPrefetcher datafetchers.Fetcher[map[uint64]passthroughdomain.PoolAPR]
+
+	poolFeesPrefetcher datafetchers.Fetcher[map[uint64]passthroughdomain.PoolFee]
 
 	logger log.Logger
 }
@@ -353,8 +355,14 @@ func (p *passthroughUseCase) GetPortfolioAssets(ctx context.Context, address str
 	return finalResult, nil
 }
 
-func (p *passthroughUseCase) RegisterAPRFetcher(aprFetcher datafetchers.Fetcher[[]passthroughdomain.PoolAPR]) {
+// RegisterAPRFetcher registers the APR fetcher for the passthrough use case.
+func (p *passthroughUseCase) RegisterAPRFetcher(aprFetcher datafetchers.Fetcher[map[uint64]passthroughdomain.PoolAPR]) {
 	p.aprPrefetcher = aprFetcher
+}
+
+// RegisterPoolFeesFetcher registers the pool fees fetcher for the passthrough use case.
+func (p *passthroughUseCase) RegisterPoolFeesFetcher(poolFeesFetcher datafetchers.Fetcher[map[uint64]passthroughdomain.PoolFee]) {
+	p.poolFeesPrefetcher = poolFeesFetcher
 }
 
 // computeCapitalizationForCoins instruments the coins with their liquiditiy capitalization values.